[Patch] Updated and documented tile based streaming
